tast-lint: Hide chromiumos/tast/diff from Tast tests.
Tast tests should use github.com/kylelemons/godebug/diff to compute
diff of two strings because it is more popular, pure-Go, and never
returns an error.

This change hides the diff package by moving it to internal/.

Note: The only usage in tast-lint continues to use the diff package
instead of github.com/kylelemons/godebug/diff because it needs to
output a diff that is accepted by patch(1).

BUG=chromium:1122465
TEST=fast_build.sh -T

Cq-Depend: chromium:2408495
Change-Id: I56c9f09988975a4a18820c74590316c24edc132c