Support concurrently running entities.
With the introduction of fixtures, it is possible that Tast CLI
observes multiple entities to run concurrently. Since currently Tast
CLI assumes that there is at most one running entity at a time, it
will abort if it observes multiple concurrently running entities.

To deal with this, this patch modifies resultsHandler to handle
concurrently running entities.

Note: This change updates Tast CLI to handle newer protocol, so we
should trigger updating Tast CLI. However fixtures are not generally
available to normal users yet, and we will make further changes to the
protocol later anyway, so we are not triggering update for now.

BUG=chromium:1035940
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run DUT meta.LocalFixture.*

Change-Id: I05a911e94a7729ffdd1be37217ed70cae1d15b57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2428406
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>