Record relative paths in results.json.
With introduction of fixtures, multiple entities with the same name
can be run multiple times, so we no longer able to locate an output
directory of an entity by "tests/$name". Instead, users should read
results.json to find outDir.

However the current results.json format is difficult to consume
because file paths in JSON are absolute and break once we copy files
around (e.g. after uploading files to cloud storage).

To solve this problem, this patch changes the output file paths
emitted in results.json to paths relative to the directory that
contains results.json.

BUG=chromium:1035940
TEST=fast_build.sh -T
TEST=fast_build.sh && ~/go/bin/tast run DUT example.Pass
TEST=cat /tmp/tast/results/latest/results.json

Change-Id: I3ce904f95b140f0d4fb8f350903435ecfab011b2