tast: hwdep: Ignore _signed suffix for model ID checking
Model ID from a device config may have _signed suffix but this is not a
part of a model name. To check model names properly in modelListed, this
suffix should be removed before checking.

BUG=b:157512297
TEST=`./fast_build.sh -vvv -t chromiumos/tast/testing/hwdep` PASSed
TEST=Run power tests on DUTs which reports _signed model ID and ensured
the parameterized tests are chosen properly based on model name.

Change-Id: I085d2473c5cc43329636b92535ef29e4677d4ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2442272
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>