Do not reset a fixture after the last test.
This patch avoids unnecessarily resetting a fixture when there is no
remaining test using the fixture.

However it is tricky to properly skip resets on traversing a fixture
tree. To make sure fixtures are reset between tests, this patch also
introduces cleanness/dirtiness of a fixture stack.

A fixture stack is clean or dirty. A stck is initially clean.
A clean stack can be marked dirty with MarkDirty. It is an error to
call MarkDirty on a dirty stack. The dirty flag can be cleared by
Reset. We call MarkDirty before running a test to make sure Reset is
called for sure between tests.

BUG=chromium:1035940
TEST=fast_build.sh -T

Change-Id: I095d3dfd8d9ec6345db063e8253dcb9f33641b92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2461067
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>