hwdep: Assume NV12 overlay support by default.
While working on a video test, I noticed that it wasn't running on a
volteer device because the SupportsNV12Overlays hwdep wasn't satisfied.
This is because findIntelSOC() didn't handle TGL correctly. Even though
that's being fixed in CL:2486095, it would be better for
SupportsNV12Overlays to be satisfied unless the SoC is known and it is
known to not support NV12 overlays. That way, we don't accidentally
prevent new devices from running certain video tests.

This CL removes the Config_SOC_UNSPECIFIED possibility from this
precondition to accomplish this.

BUG=None
TEST=video.Contents.h264_360p_hw runs on volteer

Change-Id: Ib6c7564a7d1f6064422a79a82ab2e30c3a221d20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2488757
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>