First Go CL for adding Wifi-peer machines.
This matched the router and pcap model, but allows for any number of
peers to be  added. I'll expand this, and add unit tests in my next Cl.

This is my first Go CL, so don't assume I know what I'm doing at all.

BUG=None
TEST=local unittest that doesn't mock out IPs, so tests on real
hardware. Once I figure out Mocking in Go, I'll finalize the unittest
and submit it as well.

Change-Id: I30a19f2e3aeade0e6cb15fbf5fcd2eefc37bd57b