tast: Split entity info from testing pkg.
We want to reuse the logic for extracting outDir from a Context of the
current entity, from the dut package. Those functions and structs needs
to be in a different package than testing, in order to avoid a cyclic
dependency between the two packages.

BUG=chromium:1097657
TEST=fast_build.sh -T

Change-Id: If87c2742720400901b5f62efdc4114242d584dac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2514125
Commit-Queue: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>