New subattr: firmware_smoke
Many firmware tests are useful for exercising basic library functions,
but do not need to run in the CQ. Hence, mainline is not appropriate for
these tests. It would be better to schedule them to run nightly,
alongside Tauto's existing faft_smoke suite.

This CL completes step 1 of that transition by defining a new subattr,
"firmware_smoke", under the existing "firmware" attr.

BUG=b:173440149
TEST=fast_build.sh -T

Change-Id: I43bd2f67f655f83141b2e6f802dc8c9e765e663f