tast: Introduce proto-wrapping types for better JSON customization
Refactor the code for marshalling/unmarshalling of
hardware and software feature by introducing new
wrapper types: DeviceConfigJSON & HardwareFeaturesJSON.
It also update RunTestsArgs to use JSON wrapper fields.
It is an initial CL for a larger CL which would add
sharding support to Tast runner. This CL does not
introduce any new functionality. The only difference
is that the value of DeviceConfig and HardwareFeature
in the json message would be written as "null" when
they are empty. Previously, the whole line would be
omitted.

Manual Test
2020/11/18 16:47:23 Command line: tast -verbose run -build=true 127.0.0.1:2222 ui.LauncherCreateAndRenameFolder.* ui.LauncherPinAppToShelf.*
...
2020/11/18 16:50:26 --------------------------------------------------------------------------------
2020/11/18 16:50:26 ui.LauncherCreateAndRenameFolder.clamshell_mode  [ PASS ]
2020/11/18 16:50:26 ui.LauncherCreateAndRenameFolder.tablet_mode     [ PASS ]
2020/11/18 16:50:26 ui.LauncherPinAppToShelf.clamshell_mode          [ PASS ]
2020/11/18 16:50:26 ui.LauncherPinAppToShelf.tablet_mode             [ PASS ]
2020/11/18 16:50:26 --------------------------------------------------------------------------------
2020/11/18 16:50:26 Results saved to /tmp/tast/results/20201118-164723

BUG=chromium:968262
TEST=Manual testing. Unit tests passed

Change-Id: I47d46780feff2d148b516c969dee76ec12b21655