rpc: Support OutDir in gRPC services.
This patch enables gRPC services to store files as test results.

gRPC services can obtain a temporary output directory path by
testing.ContextOutDir just as tests do. If one or more files are
saved to the directory, an extra RPC trailer is added to the gRPC
response. gRPC clients will pull the directory if it finds a RPC
trailer.

BUG=chromium:1153599
TEST=fast_build.sh -T

Change-Id: I829fe9c312f69df28adb0441756a1eced5fe8ab3