tast: Introduce embedded struct FeatureArgs
This CL introduces embedded struct FeatureArgs to
RunTestsArgs and ListTestsArgs. The feature related
arguments in RunTestsArgs were moved to the new
embedded struct. It also added FeatureArgs in the
ListTestsArgs to prepare for next CL that using
the argument in listing tests.
It is an second CL for a larger CL which would add
sharding support to Tast runner. This CL does not
introduce any new functionality.

Manual Test
2020/12/02 10:43:48 Command line: tast -verbose run -build=true octopus launcher.CreateAndRenameFolder.* launcher.PinAppToShelf.*
...
2020/12/02 10:46:56 --------------------------------------------------------------------------------
2020/12/02 10:46:56 launcher.CreateAndRenameFolder.clamshell_mode  [ PASS ]
2020/12/02 10:46:56 launcher.CreateAndRenameFolder.tablet_mode     [ PASS ]
2020/12/02 10:46:56 launcher.PinAppToShelf.clamshell_mode          [ PASS ]
2020/12/02 10:46:56 launcher.PinAppToShelf.tablet_mode             [ PASS ]
2020/12/02 10:46:56 --------------------------------------------------------------------------------
2020/12/02 10:46:56 Results saved to /tmp/tast/results/20201202-104348

BUG=chromium:968262
TEST=Manual testing. Unit tests passed

Change-Id: Ibcc7bbb064c158d8be6de694428afc9180a2838c