tast: Use EntityWithRunnabilityInfo for list test response
This CL uses EntityWithRunnabilityInfo instead of
EntityInfo list test response. The struct EntityWithRunnabilityInfo
includes EntityInfo and SkipReasons. The SkipReasons
indicate if the test will be skipped during real run.
It is an third CL for a larger CL which would add
sharding support to Tast runner. This CL does not
introduce any visible functional changes to end users,
but it change added additional field to the response to
the list test request.

Manual Test
2020/12/03 10:40:06 Command line: tast -verbose run -build=true octopus launcher.CreateAndRenameFolder.* launcher.PinAppToShelf.*
...
2020/12/03 10:43:07 --------------------------------------------------------------------------------
2020/12/03 10:43:07 launcher.CreateAndRenameFolder.clamshell_mode  [ PASS ]
2020/12/03 10:43:07 launcher.CreateAndRenameFolder.tablet_mode     [ PASS ]
2020/12/03 10:43:07 launcher.PinAppToShelf.clamshell_mode          [ PASS ]
2020/12/03 10:43:07 launcher.PinAppToShelf.tablet_mode             [ PASS ]
2020/12/03 10:43:07 --------------------------------------------------------------------------------
2020/12/03 10:43:07 Results saved to /tmp/tast/results/20201203-104006

BUG=chromium:968262
TEST=Manual testing. Unit tests passed

Change-Id: I23b177dadc6a95049c59ff4b32594c9f8212f0d7