Revert "tast-test: Disable reboot features on tast_vm"
This reverts commit 54e81ac5ca9655e60e08f86eedb5a793664187e6.

Reason for revert: After some research, this is not the root cause of LKGM failed, so we could re-enable this flag.

The tast-tests would stop the VM and restart it again between critical tasts and informational tasts, So Power.Reboot shouldn't have any affect on this issue.

And the root cause of LKGM build failed last time was clock_gettime & gettimeofday seccomp in cros-disks. (see crrev/c/2545854)

Original change's description:
> tast-test: Disable reboot features on tast_vm
>
> This is a partial revert of crrev.com/c/2518529
>
> The "reboot" feature may cause amd64-generic-full builder failed.
> And it may block chromeOS LKGM.
>
> BUG=chromium:1148527, b:170785530
> TEST=CQ
>
> Change-Id: Ic7a427993fe68800c944ed4ac3a01463eb6427db
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2545966
> Tested-by: joe Chou <yich@google.com>
> Reviewed-by: Shuhei Takahashi <nya@chromium.org>
> Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
> Reviewed-by: Andrew Moylan <amoylan@chromium.org>
> Reviewed-by: Leo Lai <cylai@google.com>
> Commit-Queue: Shuhei Takahashi <nya@chromium.org>
> Auto-Submit: joe Chou <yich@google.com>

Bug: b:170785530
Change-Id: If306a05eacc53e4dbbef7d0e8f27c7e1e441f436