tast_rtd: Create the initial version tast_rtd
This CL creates the Create initial tast_rtd which
will read invocation protobuf file and invoke tast
after it is completed. In this version, it will only
make sure the protobuf file existed and readable.

The program will exit with 0 if the file exists
and can be opened.

The program will exit with 1 if the file does not exist
or cannot be opened.

Manual Tests
$ export GOPATH=$HOME/go
$ export GOPATH=${GOPATH}:${CHROMEOS_SRC}/chroot/usr/lib/gopath
$ export GOPATH=${GOPATH}:${CHROMEOS_SRC}/src/platform/tast-tests:${CHROMEOS_SRC}/src/platform/tast
$ export GOPATH=${GOPATH}:/usr/lib64/gopath
$ go build
$ ./tast_rtd -input=nonexisting_file
2020/12/04 17:32:16 Failed to read invocation protobuf file nonexisting_file: fail to read file nonexisting_file: open nonexisting_file: no such file or directory
$ ./tast_rtd -input=test.pb
2020/12/04 17:32:23 Failed to read invocation protobuf file test.pb: fail to read file test.pb: open test.pb: permission denied
$ echo $?
1
$ chmod +r test.pb
$ ./tast_rtd -input=test.pb
$ echo $?
0

BUG=chromium:1155774
TEST=Manual testing. Unit tests passed

Change-Id: Ib5b1ed65aef8f248959dce3e1a7bd353c2eeb37a