tast: Restrict sysfs cpu vulnerabilities to boards that are green
Even though some older kernels don't report sysfs vulnerabilities we
have more boards that are failing tests relying on this software dep
because sysfs CPU reporting is incorrect. Devices that run arm32
kernels, i.e. veyron_fievel and veyron_tiger, don't report anything
because the arm32 kernel port doesn't support these sysfs files.

Furthermore, older x86 v3.18 kernels improperly report that some CPUs
are vulnerable, but that's because the detection code isn't as extensive
as newer kernels. Given that we're upreving those boards, lets skip this
test on those boards when they're running v3.18 kernels.

Similarly, arm64 kernels don't support this sysfs path in v4.4, so we
skip this test on those boards and wait for them to be upreved to a
newer kernel or for us to implement sysfs vulnerability reporting on the
v4.4 kernel if it's possible.

BUG=b:174888780, b:174889440, b:174890060, b:175345642
TEST=tast run security.CPUVulnerabilities on affected DUTs to make sure
they're skipped

Change-Id: Idfa73581fc64ef7687c3e7070ee0ddeace5f2dd3