tast: remove exception for veyron_mickey
veyron_mickey is EOL and decommissioned. Devices have been removed from
the 'wificell' pool, which is used for most tests with the
|intel_wifi_chip| label. Even if some tests use devices in other pools
and might start failing on veyron_mickey, it's a useful signal that the
decommission process is not complete and that can be fed back to the lab
team.

BUG=b:173723099
TEST=None

Change-Id: I8bf26b89d2328f76ff4a158c4c876561c6dd6d3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2595202
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>