testing: relax variable access restriction
We were posing restriction that variables in the form of foo.Bar
are only accessible from tests in the foo category.
This restriciton turned out to be too tight; for example it disallows
creating preconditions that use secrets.

This change relaxes the restriction, allowing all tests to access
variables in the form of foo.Bar.

BUG=chromium:1157754
TEST=fast_build.sh -T

Change-Id: I4144ce6d0b791288c13eee334355d564c41343d4