rpc: Conditionally initialize user-defined gRPC services
This is a part of series of patches to refactor the rpc package.

This patch adds the need_user_services field to the handshake request
message. User-defined gRPC services are initialized only when this
field is set to true.

While this patch changes the protocol between remote tests and local
services, it has no compatibility implication because remote test
bundles and local test bundles are always in sync.

BUG=chromium:1128259
TEST=fast_build.sh -T
TEST=tast run DUT example.GRPC

Change-Id: I9ac985a52fe2ade071ce925c9089a7182d8fe061