rpc: Convert per-call hooks to the callback style
This is a part of series of patches to refactor the rpc package.

gRPC client/server provided by the rpc package installs per-call hooks
to perform some operations before/after gRPC calls. This patch rewrites
them to the callback style so that the code flow is clearer.

This is a pure refactoring change. No functional change is expected.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I68f0c39f220a4173366d6ef1202cecef971e437b