rpc: Allow registering extra core services
This is a part of series of patches to refactor the rpc package.

This patch adds a new callback argument to rpc.RunServer which can be
used to install extra gRPC services.

rpc.RunServer still takes []*testing.Service and install those services
internally. It is not easy to do this externally in the new callback
because Context passed to user-defined gRPC service initialization is
tied to remoteLoggingServer which has to be initialized before
grpc.Server.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Id25b58069a0b9c99bf050449a87f2d171cbe5fe6