Define fingerprint MCU board specific attrs
This is a correction of http://crrev/c/2593152. At the time we were
defining "fingerprint-mcu" attr and hoping to use a custom swarming
label "fingerprint-dev-board=dragonclaw/icetower" in suite scheduler
to pick the correct setup. However, we learned that managing the
swarming label is complex, and it's easier to just define separate
pools for the two setups.

BUG=b:158580909
TEST=none

Change-Id: I62964cb5e3cc0062e07c3cc7b313d5c86978ac37