Tast docs: URLs point to main instead of master
Tast documentation contains a lot of URLs to code files, containing the
"master" branch name in the URL. However, in December several
repositories switched their main branch from "master" to "main". As a
result, the "master" URLs now point to a frozen codepoint.

This URL switches the URLs in tast/docs/ to the "main" branch instead.

BUG=b:176813234
TEST=Open docs in Gitiles, click all URLs

Change-Id: I0dd078c46a4bbc4b62ea4d9cec260897ee4178c3