tast: Avoid writing fixture results on abort
When the test runner is aborted, we call streamedResultsWriter.write
unconditionally for all running entities. We should not call it for
fixtures so that streamed_results.jsonl does not contain any fixture
results.

To make the matter worse, we call streamedResultsWriter.write with
update=true in this case. This means that, if the test runner aborts
during fixture execution, the last test result in streamed_results.json
can be lost.

BUG=chromium:1163765
BUG=b:176931008
TEST=fast_build.sh -T

Change-Id: If88558dabd7db5e44caf8a4ec15b662d418ab9d9