bundle: Consolidate local/remote delegate structs
We used to have separate logic to handle local tests and remote tests
in general. But this approach turned out to result in many code
duplication and it's getting more difficult to maintain.

This patch introduces the bundle.Delegate struct which is a union of
former LocalDelegate and RemoteDelegate, so that we don't need to
handle them separately.

BUG=chromium:1134060
TEST=fast_build.sh -T

Change-Id: Ibd3071674b1f83fb6d7763b525957828f3e5fca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2627227
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>