hwdep: Add check for SOCs supporting 30bpp formats
This change adds a function that indicates whether a given SOC supports
30bpp framebuffer formats, which is true for Intel SOCs Kabylake on, and
AMD SOCs Zork on, but not for ARM SOCs.

This CL is a prerequisite for adding a Tast test to exercise the
null_platform_test for 30bpp formats.

BUG=b:169781466
TEST=Added dep to a 30bpp version of the NullPlatformFormat test (found
here
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/graphics/null_platform_format.go;l=30;drc=fe346aa73ba672794d62ce8d578ec4d3d06f41d0)
and verified the test ran on my device with tast run <DEVICE>@<LOCATION> graphics.NullPlatformFormat.30bpp

Change-Id: I98b15985a472b6917b267fce2c22b2f3b488e37f