hwdep: Add check for SOCs supporting 30bpp formats
This change adds a function that indicates whether a given SOC supports
30bpp framebuffer formats, which is true for Intel SOCs Kabylake on, and
AMD SOCs Zork on, but not for ARM SOCs.

This CL is a prerequisite for adding a Tast test to exercise the
null_platform_test for 30bpp formats.

BUG=b:169781466
TEST=Added dep to a 30bpp version of the NullPlatformFormat test (with
crrev.com/c/2643960 patched) and verified the test ran on my device with tast run <DEVICE>@<LOCATION> graphics.NullPlatformFormat.30bpp

Change-Id: I98b15985a472b6917b267fce2c22b2f3b488e37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2641001
Tested-by: Clarissa Garvey <clarissagarvey@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>