go/tast-running: Add section about servo
Servo is used for several Tast tests. Running a Tast test with Servo has
a few quirks: it must be specified in the command-line, and servod must
be initiated separately. This CL adds a section to go/tast-running with
instructions on how to run a Tast test with servo.

BUG=None
TEST=Open in gitiles

Change-Id: I2cfad72c152997cb80929094fd624cdd99047663
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2653861
Tested-by: Greg Edelston <gredelston@google.com>
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>