logging: Export a struct type instead of an interface
Due to removal of the -fancy flag, simpleLogger is now the only
implementation of the Logger interface. To avoid the unnecessary
abstraction, this patch replaces the Logger interface with the
former simpleLogger struct type.

This is a refactoring change. No functional change is expected.

BUG=chromium:1174982
BUG=chromium:989733
TEST=fast_build.sh -T

Change-Id: I1857754f30f5e6f8e9772e0f1f1c1fc7cd4678d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2671724
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>