test: Inherit build= flag from parent for meta tests
There are meta tests that run a tast command with another test.
Those child tests should be:

When running locally, developer may want to build tests.
So build=true or build=false may be used.
It would be better to build child tests as well when the parent test is
built.

When running in the lab, tests are already installed.
So build=false for both the parent and child tests.

Considering these usage, make tast to inherit the -build flag value to
the child tests.

BUG=chromium:1153263
TEST=fast_build.sh -t chromiumos/tast/cmd/tast/internal/run

Change-Id: Ie27ce3ab8399113a4d43bd330b1984399cadd4f3