test: Inherit build= flag from outer test, for meta tests
There are meta tests that run a tast command with another test.

1. When running in the lab, tests are already installed.
So build=false for both the inner and outer tests.

2. When running locally, developer may want to build tests.
So both build=true and build=false can be used for outer test.
When build=true, building the inner tests as well is more consistent.
Also, the latest data files for inner tests should be pushed.

Considering these usage, make tast to inherit the -build flag value to
the inner tests.

BUG=chromium:1153263
TEST=fast_build.sh -t chromiumos/tast/cmd/tast/internal/run
TEST=manually, as noted in the bug

Change-Id: Ie27ce3ab8399113a4d43bd330b1984399cadd4f3