Tast: Add WifiIntel hardware dependency
This CL adds the hardware dependency WifiIntel to Tast, which allows
tests to run iff they use Intel WiFi.

This is done by defining a list of devices which do not use Intel WiFi
and skipping on those devices. This is a temporary measure which will
be replaced when WiFi configs are added to device.Config as discussed in
${BUG}.

Some platforms have models both with and without Intel WiFi, so a
combination of SkipOnPlatform and SkipOnModel is used to account for
this. This also prevents us from having to maintain an exhaustive list
of all models without Intel WiFi, which would be very long and
difficult to maintain.

BUG=b:180133528
TEST=fast_build.sh -T
TEST=Add `HardwareDeps: hwdep.D(hwdep.WifiIntel())` to a test and run on
various boards.
Results:
zork(morphius): Runs correctly
zork(ezkinil): Skips "ModelId matched with skip-on list"
grunt-kernelnext(careena): Skips "PlatformId matched with skip-on list"

Change-Id: Ie22eb7617cdd466b218eca0fddf8da63ba60de72