run: Move build-related constants to the build package
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

In upcoming changes we want to move config.go to a separate package.
Thus config.go should not refer constants defined in other files to
avoid circular imports.

This patch moves a few build-related constants to internal/build.

This is a pure refactoring change; no functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: I81646d53a56e741f1510d4ef060cd87b5d686a70