docs: mention special cases for group:wificell
Some WiFi-specific tests do not actually require to be run in a
wificell. For example, smoke tests that verify that the WLAN interface
interface has been created at boot do not necessarily need to connect to
an AP. However, we still prefer those tests to also declare the
"wificell" requirement so they can all be part of the same
"wifi_matfunc" WiFi validation suite. That reduces confusion, in
particular for partners.

BUG=b:181060056
TEST=preview in gitiles

Change-Id: Ia03fd444547b9443dde06e264ffe7b4f8ef2c06d