run: Attach closeEphemeralDevserver to State
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

In upcoming changes we want to move config.go to a separate package.
Currently State.Close calls closeEphemeralDevserver, which would
introduce circular imports on moving config.go.

This patch attaches closeEphemeralDevserver to State.

This is a pure refactoring change; no functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: I89c55b1f1d8a40f1b762859e496db8ee58e1ae06