Support IPv6-only host systems
On systems where IPv4 is disabled, even the IPv4 loopback address
127.0.0.1 is unavailable. In order to support such systems we have
to specify "localhost" on listening on the host so that an
appropriate loopback address is chosen by the system.

Note that we still use IPv4 on listening on the DUT by SSH reverse
port forwarding. As far as I know Chrome OS does not provide a way to
disable IPv4 on the loopback device today.

BUG=chromium:1179040
TEST=fast_build.sh -T
TEST=fast_build.sh && ~/go/bin/tast run $DUT meta.RunTests.files_remote

Change-Id: Id52b101b8b5f2d388224e25aa66c6706166848d3