run: Export symbols in config.go
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

In upcoming changes we want to move config.go to a separate package.
Then other files in internal/run will lose access to unexported
symbols.

This patch exports struct fields, types and methods that are referred
outside config.go to prepare for the move.

This is a pure renaming change. No functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: Id9c1a06e79a8af6001e60d6a331d4bcdb2731a95