run: Move EntityResult to a new package
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

In upcoming changes we want to move config.go to a separate package.
This patch moves the definition of EntityResult (and EntityError) to
a new package so that Config.TestsToRun can refer EntityResult after
the move.

This is a pure code move change. No functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: I70bad73c49f7e5acd4c2cdc6b7176e1129638297