run: Remove direct access to State.Hst
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

State.Hst is a cached connection to the DUT. It should be accessed
only by connectToTarget and State.Close. This patch removes
unexpected direct access to State.Hst.

This is a refactoring change. No functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: Ibbe6493a77307c3b0ee054e61ebddbf9af8fbe16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2751491
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>