run: Move readBootID to the linuxssh package
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

In the next patch, we want to call readBootID from multiple packages.
Thus this patch moves the function to the linuxssh package.

This is a pure code move change. No functional change is expected.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: Ic97c20b34b9339258c5a47a7b815fa355202fb8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2751492
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>