run: Extract target connection cache from State
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

Currently State holds a cached connection to the target, and it is
managed by connectToTarget and State.Close. Since connectToTarget
is defined in the run package, any function in the run package that
calls connectToTarget cannot be separated out to a subpackage without
introducing circular dependencies.

This patch introduces a new package run/target which is responsible
for maintaining resources related to targets. target.ConnCache holds
a cached connection to the target. target.ConnCache should be passed
explicitly in function arguments instead of implicitly passing it via
State, which makes it clear which function may attempt to reconnect.

This patch mostly does not change the current behavior, but only one
exception is that we no longer reuse a cached connection on collecting
sysinfo at the end of testing; that is, collectSysInfo creates a new
ConnCache instead of taking one in the parameter. This is because we
cannot pass target.ConnCache constructed in run.Run to run.WriteResults.
Eventually I want to change Run to call WriteResults directly to allow
sharing data between the two phase without relying on State, but I'm
not making the change for now because it will be a lot of work.

BUG=chromium:1179179
TEST=fast_build.sh -T
TEST=fast_build.sh && ~/go/bin/tast run $DUT example.Pass

Change-Id: I3e433a7a6d91c75d7c981aaea8330ecfa199184f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2751493
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>