run: Randomize ephemeral devserver port number
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

Using a fixed port number is generally bad because it can lead to
resource conflicts.

In fact, we have to stop the ephemeral devserver before running
remote tests to avoid possible conflicts with the ephemeral devserver
started by meta remote tests.

This has a design implication that we cannot assume that ephemeral
devserver lifetime is not always equal to an SSH connection to the
target device, making upcoming refactoring harder.

Thus this patch makes the ephemeral devserver stop using a fixed
port number, and remove a hack to stop the ephemeral devserver early.

BUG=chromium:1179179
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run $DUT meta.DisruptSSH meta.RunTests.faillog

Cq-Depend: chromium:2772381
Change-Id: I25fd82bf157e50d8e9b517e7cfc76f15139c30d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2751503
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>