run: Extract fake test runners to a subpackage
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

Extra fake test runners in unit test to a subpackage. This allows
extracting more components in the run package to subpackages when
their unit tests use fake test runners.

This is a unit test only change. No change to production binary.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: If76c2baa62ece8fd4afc93f3b44dea41fec01cab