run: Move code around within the run package
This is a part of changes to split internal/run into multiple
subpackages (crbug.com/1179179).

This patch moves several functions and their associated tests within
the run package so that diff is minimized for upcoming changes that
extracts those code to separate subpackages.

This is a large change, but it is really a pure code move.

BUG=chromium:1179179
TEST=fast_build.sh -T

Change-Id: I45f379da6b19175ce01e0eb84336ae7587037dab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2759315
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>