planner: Panic on calling top for an empty fixture stack
FixtureStack.top used to return nil when a fixture stack is empty.
However calling top for an empty fixture stack is usually wrong, so
returning nil might hide the issue.

This patch updates top to panic when it is called for an empty fixture
stack instead.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I327e0aa200cb1107a0fb2955b932f666aad3e029
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2777976
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>