protocol: Rename Error to ErrorReport
In the next change we're introducing "Error" message which is used in
the protocol between Tast CLI and test bundles. Since the usage of
the new message is more popular than the current "Error" message,
renaming existing message to ErrorReport.

This is a pure rename change.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Iae2c1698c6d023d70c9b0720bb2f5885aa887e35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2777981
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>