bundle: Rename runConfig to staticConfig
Currently there are two "configuration" structs given to a bundle:
bundle.Args and bundle.runConfig. In a sense both of the two are used
to customize the run-time behavior, so it is not meaningful to name
the latter struct "runConfig".

This patch renames runConfig to staticConfig because the notable
difference between Args and runConfig is that the former is supplied
by the user at run-time while the latter is supplied statically from
test bundle's main functions.

This patch also updates the comments for staticConfig to reflect the
reality today.

This is a pure renaming change.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Iac5738c5cf8ad3b1fd0bbc82b566be01a42f0735
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2777977
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>