Move JSON protocol structs to jsonprotocol package
This change moves following structs in testing package that are part
of Tast JSON protocol:

- Error
- EntityType
- EntityInfo
- EntityWithRunnabilityInfo

This change makes it clear which components have dependency to JSON
protocol. To accomplish gRPC protocol migration we have to minimize
imports of jsonprotocol package.

This is a pure code move.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I818ee7e47011d1e018b93aab513119df95f20bd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2777980
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>