crash: lift the limit on the number of crash dumps to save
Tast currently copies up to 3 crash dumps per run. This was sufficient
when the logic was introduced years ago, but today we have a lot of
Tast tests and may cause multiple crashes per run.

This patch lifts the limit, but instead adds a logic to skip copying
core dumps (.core). Core dumps are too large to copy. Minidumps (.dmp)
should be sufficient.

BUG=chromium:1183458
TEST=fast_build.sh -T

Change-Id: I3663eed2dcbf028398825a56e5674e31ffad7183