Consistently report dependency check errors as test errors
Currently a test can be skipped while reporting errors, e.g. when it
depends on undefined software features. However this semantics is very
confusing for test result consumers, and it is likely they are simply
treated as skips. This is not favorable because run time errors are
silently ignored by test skips.

This patch fixes the issue by not marking a test as skipped when run
time dependency checks for the test has failed. To better enforce
the new semantics, this patch also updates signatures of several related
functions/types.

BUG=chromium:1194851
BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Ic3bbbde3e7c2a1b1ba0a655933119b162d48a08a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2799391
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>