runner: Use gRPC to list entities in bundles
Finally, this is the very first patch in the series to make a meaningful
protocol change. We start it by using gRPC between test runners and
test bundles on getting a list of entities

Though this patch introduces a protocol change, it is between test
runners and test bundles which are guaranteed to be built from the
same snapshot of source codes. Thus we don't need to care protocol
compatibility issues.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I1b2a95f89f8bd570cb305aa0bdaa3a8933214031